refactor: move drag sequence simulations into separate module
