+++ packages/test-utils/src/index.ts
@@ -1,2 +1,3 @@
 export * from './wrappers'
 export * from './utils'
+export * from './eventSimulation'
